Cambios varios en facturacion y consumibles

+++ FrbaHotel/FrbaHotel/ClasesTabla/Consumible.cs
@@ -9,7 +9,7 @@ namespace FrbaHotel.ClasesTabla
     {
         public decimal codigo { get; set; }
         public string descripcion { get; set; }
-        public int cantidad { get; set; }
+        public decimal cantidad { get; set; }
         public decimal precio { get; set; }
     }
 }
